+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - [Resources] Functions in the `resources` module no longer have default values for the `version` argument. [#297]
 
+- [Compatibility] Drop support for Python 2.7. [#311]
 - [Compatibility] Drop support for Python 3.4. [#309]
 
 ### Fixed